Fixed a massive memory leak due to debug draw
The debug draw variables were being re-generated every time debug draw was called, instead of being reused.
Changed entity class so that instead of draw debug functon, it now has a m_bEnableDebugDraw bool, and debug draw is a part of the entity's draw function. Also added a virtual function, InitializeDebugDraw() so that every entity must specify at least that function (does not have to be filled in). This is due to the fact that debug draws for spheres are different to boxes, and must be set differently, until a more intuitive way is created to do debug draw